* moves Spinner into new module Common * reorganizes styles
