+++ app/javascript/src/Applications/components/DefaultPlanSelect.scss
@@ -1,0 +1,8 @@
+.pf-c-select__toggle.pf-m-disabled input {
+  background-color: var(--pf-global--disabled-color--300);
+}
+
+.pf-c-select__menu {
+  max-height: 150px;
+  overflow-y: scroll;
+}
