+++ app/javascript/src/Applications/components/DefaultPlanSelectCard.scss
@@ -1,14 +1,5 @@
 @import "~@patternfly/patternfly/components/Card/card.css";
 
-#default-plan-select-card {
+.pf-c-card#default_plan_card {
   margin-bottom: var(--pf-c-page__main-section--PaddingBottom);
-
-  .pf-c-select__toggle.pf-m-disabled input {
-    background-color: var(--pf-global--disabled-color--300);
-  }
-
-  .pf-c-select__menu {
-    max-height: 150px;
-    overflow-y: scroll;
-  }
 }
