+++ app/javascript/src/Applications/index.js
@@ -3,4 +3,3 @@
 export * from 'Applications/components/DefaultPlanSelect'
 export * from 'Applications/components/DefaultPlanSelectCard'
 export * from 'Applications/components/DefaultPlanSelectWrapper'
-export * from 'Applications/components/MySpinner'
