+++ app/javascript/src/Common/components/Spinner.scss
@@ -1,4 +1,4 @@
-// Spinner Component
+// The Spinner is still not available in the version of @patternfly/react-core we use, therefor we need all the following styles
 
 .pf-c-spinner {
   --pf-c-spinner--AnimationDuration: 1.5s;
