+++ app/javascript/src/Common/index.jsx
@@ -1,0 +1,3 @@
+// @flow
+
+export * from './components/Spinner'
