--- app/views/api/plans/_default_plan.html.slim
@@ -1,2 +1,0 @@
-div id='default_plan' data=default_plan_data(@service, @plans)
-  = javascript_pack_tag 'default_plan_selector'
