+++ app/views/api/plans/_listing.html.slim
@@ -1,6 +1,7 @@
 section class="pf-c-page__main-section pf-m-no-padding"
   - if @plans.not_custom.size > 0
-    = render partial: 'api/plans/default_plan'
+    div id='default_plan' data=default_plan_data(@service, @plans)
+      = javascript_pack_tag 'default_plan_selector'
 
   div.pf-c-card
     / TODO: use card body and other pf styles instead of .data class
